+++ cmd/kogito/command/converter/runtimetype_converter_test.go
@@ -16,17 +16,76 @@ package converter
 
 import (
 	"github.com/kiegroup/kogito-cloud-operator/cmd/kogito/command/flag"
+	"github.com/kiegroup/kogito-cloud-operator/cmd/kogito/command/test"
 	"github.com/kiegroup/kogito-cloud-operator/pkg/apis/app/v1alpha1"
 	"github.com/stretchr/testify/assert"
+	"os"
 	"testing"
 )
 
-func Test_FromRuntimeFlagsToRuntimeType(t *testing.T) {
+func Test_FromRuntimeFlagsToRuntimeType_SpringBoot(t *testing.T) {
 	flags := &flag.RuntimeTypeFlags{
 		Runtime: "springboot",
 	}
 
 	runtimeType := FromRuntimeFlagsToRuntimeType(flags)
-	assert.NotNil(t, runtimeType)
 	assert.Equal(t, v1alpha1.SpringBootRuntimeType, runtimeType)
+}
+
+func Test_FromRuntimeFlagsToRuntimeType_Quarkus(t *testing.T) {
+	flags := &flag.RuntimeTypeFlags{
+		Runtime: "quarkus",
+	}
+
+	runtimeType := FromRuntimeFlagsToRuntimeType(flags)
+	assert.Equal(t, v1alpha1.QuarkusRuntimeType, runtimeType)
+}
+
+func Test_FromArgsToRuntime_NonBinaryBuild(t *testing.T) {
+	flags := &flag.RuntimeTypeFlags{
+		Runtime: "quarkus",
+	}
+
+	runtimeType, err := FromArgsToRuntimeType(flags, flag.GitRepositoryResource, "https://github.com/kiegroup/kogito-examples/blob/stable/process-scripts-quarkus/src/main/resources/org/acme/travels/scripts.bpmn")
+	assert.Nil(t, err)
+	assert.Equal(t, v1alpha1.QuarkusRuntimeType, runtimeType)
+}
+
+func Test_FromArgsToRuntime_BinaryBuild_SpringBoot(t *testing.T) {
+	flags := &flag.RuntimeTypeFlags{
+		Runtime: "quarkus",
+	}
+
+	tmpDir := test.TempDirWithFile("target", "*-runner-*.jar")
+	defer os.RemoveAll(tmpDir)
+
+	runtimeType, err := FromArgsToRuntimeType(flags, flag.LocalBinaryDirectoryResource, tmpDir)
+	assert.Nil(t, err)
+	assert.Equal(t, v1alpha1.SpringBootRuntimeType, runtimeType)
+}
+
+func Test_FromArgsToRuntime_BinaryBuild_Quarkus(t *testing.T) {
+	flags := &flag.RuntimeTypeFlags{
+		Runtime: "quarkus",
+	}
+
+	tmpDir := test.TempDirWithFile("target", "*-runner.jar")
+	defer os.RemoveAll(tmpDir)
+
+	runtimeType, err := FromArgsToRuntimeType(flags, flag.LocalBinaryDirectoryResource, tmpDir)
+	assert.Nil(t, err)
+	assert.Equal(t, v1alpha1.QuarkusRuntimeType, runtimeType)
+}
+
+func Test_FromArgsToRuntime_BinaryBuild_QuarkusNative(t *testing.T) {
+	flags := &flag.RuntimeTypeFlags{
+		Runtime: "quarkus",
+	}
+
+	tmpDir := test.TempDirWithFile("target", "*-runner")
+	defer os.RemoveAll(tmpDir)
+
+	runtimeType, err := FromArgsToRuntimeType(flags, flag.LocalBinaryDirectoryResource, tmpDir)
+	assert.Nil(t, err)
+	assert.Equal(t, v1alpha1.QuarkusRuntimeType, runtimeType)
 }