+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 - [KOGITO-3270](https://issues.redhat.com/browse/KOGITO-3270) - Operator should support ConfigMap injecting in KogitoServices
 - [KOGITO-2905](https://issues.redhat.com/browse/KOGITO-2905) - Change in image definition in the CRDs
+- [KOGITO-2113](https://issues.redhat.com/browse/KOGITO-2113) - Simplify binary build deployment
 ## Bug Fixes
 
 - [KOGITO-3382](https://issues.redhat.com/browse/KOGITO-3382) - Transitive dependency of Keycloak operator module missing
-